🚨 Extract `token_storage` from `models_config`
TG-2031
